+++ src/app/scroll-feed/page.tsx
@@ -1,11 +1,11 @@
 export const dynamic = "force-dynamic";
 
-import RedditPostStack from "@/components/reddit-post-stack";
+import ScrollFeed from "@/components/scroll-feed";
 
-export default function Page2() {
+export default function Page1() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-center bg-gray-700 text-white">
-      <RedditPostStack />
+    <main className="flex min-h-screen flex-col items-center justify-center bg-gray-800 text-white">
+      <ScrollFeed />
     </main>
   );
 }
