+++ src/app/wheel-feed/page.tsx
@@ -1,9 +1,11 @@
-import ScrollFeed from "@/components/scroll-feed";
+export const dynamic = "force-dynamic";
 
-export default function Page1() {
+import RedditPostStack from "@/components/reddit-post-stack";
+
+export default function Page2() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-center bg-gray-800 text-white">
-      <ScrollFeed />
+    <main className="flex min-h-screen flex-col items-center justify-center bg-gray-700 text-white">
+      <RedditPostStack />
     </main>
   );
 }
